model explain bug
