--- streamingpro-mlsql/src/main/java/streaming/dsl/load/batch/ModelSource.scala
@@ -1,10 +1,0 @@
-package streaming.dsl.load.batch
-
-/**
-  * Created by allwefantasy on 12/9/2018.
-  */
-class ModelSource {
-  def output() = {
-
-  }
-}
