+++ streamingpro-mlsql/src/test/scala/streaming/test/dsl/DslSpec.scala
@@ -593,6 +593,9 @@ class DslSpec extends BasicSparkOperation with SpecFunctions with BasicMLSQLConf
       spark.sql("select * from output").show()
       ScriptSQLExec.parse("load workflow.`` as output;", ssel)
       spark.sql("select * from output").show()
+
+      ScriptSQLExec.parse("load workflow.`types` as output;", ssel)
+      spark.sql("select * from output").show()
     }
   }
